+++ src/gtk/CairoImageConv.cpp
@@ -18,10 +18,11 @@ using LibRpTexture::rp_image;
 
 /**
  * Convert an rp_image to cairo_surface_t.
- * @param img	[in] rp_image.
+ * @param img		[in] rp_image.
+ * @param premultiply	[in] If true, premultiply. Needed for display; NOT needed for PNG.
  * @return GdkPixbuf, or nullptr on error.
  */
-cairo_surface_t *CairoImageConv::rp_image_to_cairo_surface_t(const rp_image *img)
+cairo_surface_t *CairoImageConv::rp_image_to_cairo_surface_t(const rp_image *img, bool premultiply)
 {
 	assert(img != nullptr);
 	if (unlikely(!img || !img->isValid()))
@@ -46,10 +47,16 @@ cairo_surface_t *CairoImageConv::rp_image_to_cairo_surface_t(const rp_image *img
 
 	switch (img->format()) {
 		case rp_image::FORMAT_ARGB32: {
-			// Premultiply the image first.
-			// TODO: Combined dup()/premultiply() function?
-			unique_ptr<rp_image> img_prex(img->dup());
-			img_prex->premultiply();
+			const rp_image *img_prex;
+			if (premultiply) {
+				// Premultiply the image first.
+				// TODO: Combined dup()/premultiply() function?
+				img_prex = img->dup();
+				const_cast<rp_image*>(img_prex)->premultiply();
+			} else {
+				// No premultiplication.
+				img_prex = img;
+			}
 
 			// Copy the image data.
 			int dest_stride = cairo_image_surface_get_stride(surface);
@@ -78,11 +85,14 @@ cairo_surface_t *CairoImageConv::rp_image_to_cairo_surface_t(const rp_image *img
 
 			// Mark the surface as dirty.
 			cairo_surface_mark_dirty(surface);
+			if (premultiply) {
+				delete img_prex;
+			}
 			break;
 		}
 
 		case rp_image::FORMAT_CI8: {
-			const uint32_t *palette = img->palette();
+			const uint32_t *const palette = img->palette();
 			const int palette_len = img->palette_len();
 			assert(palette != nullptr);
 			assert(palette_len > 0);
@@ -92,12 +102,18 @@ cairo_surface_t *CairoImageConv::rp_image_to_cairo_surface_t(const rp_image *img
 
 			// Premultiply the palette.
 			std::array<uint32_t, 256> pal_prex;
-			for (int i = 0; i < palette_len; i++) {
-				pal_prex[i] = rp_image::premultiply_pixel(palette[i]);
-			}
-			if (palette_len < (int)pal_prex.size()) {
-				// Clear the rest of the palette.
-				memset(&pal_prex[palette_len], 0, (pal_prex.size() - palette_len) * sizeof(uint32_t));
+			const uint32_t *pal_toUse;
+			if (premultiply) {
+				for (int i = 0; i < palette_len; i++) {
+					pal_prex[i] = rp_image::premultiply_pixel(palette[i]);
+				}
+				if (palette_len < (int)pal_prex.size()) {
+					// Clear the rest of the palette.
+					memset(&pal_prex[palette_len], 0, (pal_prex.size() - palette_len) * sizeof(uint32_t));
+				}
+				pal_toUse = pal_prex.data();
+			} else {
+				pal_toUse = palette;
 			}
 
 			// Copy the image data.
@@ -108,16 +124,16 @@ cairo_surface_t *CairoImageConv::rp_image_to_cairo_surface_t(const rp_image *img
 			for (unsigned int y = (unsigned int)height; y > 0; y--) {
 				unsigned int x;
 				for (x = (unsigned int)width; x > 3; x -= 4) {
-					px_dest[0] = pal_prex[img_buf[0]];
-					px_dest[1] = pal_prex[img_buf[1]];
-					px_dest[2] = pal_prex[img_buf[2]];
-					px_dest[3] = pal_prex[img_buf[3]];
+					px_dest[0] = pal_toUse[img_buf[0]];
+					px_dest[1] = pal_toUse[img_buf[1]];
+					px_dest[2] = pal_toUse[img_buf[2]];
+					px_dest[3] = pal_toUse[img_buf[3]];
 					px_dest += 4;
 					img_buf += 4;
 				}
 				for (; x > 0; x--, px_dest++, img_buf++) {
 					// Last pixels.
-					*px_dest = pal_prex[*img_buf];
+					*px_dest = pal_toUse[*img_buf];
 					px_dest++;
 					img_buf++;
 				}