+++ src/gtk/CairoImageConv.hpp
@@ -30,10 +30,11 @@ class CairoImageConv
 	public:
 		/**
 		 * Convert an rp_image to cairo_surface_t.
-		 * @param img	[in] rp_image.
+		 * @param img		[in] rp_image.
+		 * @param premultiply	[in] If true, premultiply. Needed for display; NOT needed for PNG.
 		 * @return cairo_surface_t, or nullptr on error.
 		 */
-		static cairo_surface_t *rp_image_to_cairo_surface_t(const LibRpTexture::rp_image *img);
+		static cairo_surface_t *rp_image_to_cairo_surface_t(const LibRpTexture::rp_image *img, bool premultiply = true);
 };
 
 #endif /* __ROMPROPERTIES_GTK_CAIROIMAGECONV_HPP__ */